added staking types
