+++ src/plugins/cosmos/components/ClaimButton/ClaimButton.tsx
@@ -1,7 +1,7 @@
 import { FlexProps } from '@chakra-ui/layout'
 import { Button } from '@chakra-ui/react'
 import { CAIP19 } from '@shapeshiftoss/caip'
-import { StakingAction } from 'plugins/cosmos/components/modals/Staking/Staking'
+import { StakingAction } from 'plugins/cosmos/components/modals/Staking/StakingTypes'
 import { Text } from 'components/Text'
 import { useModal } from 'context/ModalProvider/ModalProvider'
