+++ src/plugins/cosmos/components/modals/Staking/StakingTypes.ts
@@ -1,0 +1,35 @@
+import { CAIP19 } from '@shapeshiftoss/caip'
+
+export enum StakingAction {
+  Stake = 'stake',
+  Unstake = 'unstake',
+  Overview = 'overview',
+  Claim = 'claim'
+}
+
+export type StakingModalProps = {
+  assetId: CAIP19
+  action: StakingAction
+}
+
+export enum StakeRoutes {
+  Stake = '/stake',
+  Unstake = '/unstake',
+  StakeConfirm = '/stake/confirm',
+  UnstakeConfirm = '/unstake/confirm',
+  Overview = '/stake/overview',
+  Claim = '/claim'
+}
+
+export const entries = [
+  StakeRoutes.Stake,
+  StakeRoutes.Unstake,
+  StakeRoutes.StakeConfirm,
+  StakeRoutes.UnstakeConfirm,
+  StakeRoutes.Overview
+]
+
+export enum UnstakingPath {
+  Confirm = '/unstaking/confirm',
+  Broadcast = '/unstaking/broadcast'
+}
