+++ src/plugins/cosmos/components/modals/Staking/views/UnstakeConfirmRouter.tsx
@@ -7,7 +7,7 @@ import { RouteSteps } from 'components/RouteSteps/RouteSteps'
 import { SlideTransition } from 'components/SlideTransition'
 import { BigNumber } from 'lib/bignumber/bignumber'
 
-import { StakingAction } from '../Staking'
+import { StakingAction, UnstakingPath } from '../StakingTypes'
 import { UnstakeConfirm } from './UnstakeConfirm'
 
 type UnstakingConfirmProps = {
@@ -15,11 +15,6 @@ type UnstakingConfirmProps = {
   assetId: CAIP19
   fiatRate: BigNumber
   onCancel: () => void
-}
-
-export enum UnstakingPath {
-  Confirm = '/unstaking/confirm',
-  Broadcast = '/unstaking/broadcast'
 }
 
 export const withdrawRoutes = [
